fixing auth problems
